feat: change signature for "register" UDF to fix error handling
When using the 'register' function directly with `q.Call('register')`,
you must now pass in an email argument before the data object, like this:
    ```
    q.Call('register', password, email, {
         confirmedEmail: false,
         details,
         email,
         locale,
         username,
     }),
    ```
Previously, we sourced the email from within the data object, but
this was the cause of a bug. If you are using the `register` utility
imported from the `faunauth` npm package, you can ignore this change.
In other words, if you only do this:

    ```
    import { register } from 'faunauth';

    // later, call register() somewhere in your code
    ```
then you can ignore this change.